Extract light device setup module
